Wording improvements
Wording adjusted to be more imperative - it's more suitable for a HowTo guide this way. All the "let's" and "can" are usually reserved for tutorials.
Link text changed to not have "here".
Added a link to Terraform in its introduction.
Line splits - splitting longer lines to limit the scope of changes in git commits later.
Removed redundant empty lines.
Other minor improvements.